add docstrings, fix missing slashes in home path of labelmaker

+++ cloudmesh/cc/job/localhost/Job.py
@@ -353,7 +353,7 @@ class Job:
         specified in the period in seconds,
         till the job has completed
         :param period: time in seconds to check, as an interval
-        :type period: int
+        :type period: float
         :return: does not return anything
         :rtype: None
         """
@@ -388,7 +388,7 @@ class Job:
         """
         kills the job
         :param period: interval to use for waiting for log/pid
-        :type period: int
+        :type period: float
         :returns:
             - pid - process identifier as str
             - child - child process as str
