+++ cloudmesh/cc/job/ssh/Job.py
@@ -274,7 +274,7 @@ class Job:
         specified in the period in seconds,
         till the job has completed
         :param period: time in seconds to check, as an interval
-        :type period: int
+        :type period: float
         :return: does not return anything
         :rtype: None
         """
@@ -309,7 +309,7 @@ class Job:
         """
         kills the job
         :param period: interval to use for waiting for log/pid
-        :type period: int
+        :type period: float
         :returns:
             - pid - process id of the script
             - child - child of the script
